+++ app/src/components/market/sections/common_sections/messages/market_resolution_message/img/Won.tsx
