+++ app/src/components/market/sections/common_sections/messages/token_item/index.tsx
