+++ app/src/components/market/sections/common_sections/messages/warning_message/img/Alert.tsx
