+++ app/src/components/market/sections/market_create/steps/items/ask_question_step.tsx
@@ -16,8 +16,8 @@ import { QuestionInput } from '../../../../../common/form/question_input'
 import { Categories } from '../../../../common/categories'
 import { MarginsButton } from '../../../../common/common_styled'
 import { CreateCard } from '../../../../common/create_card'
-import { WarningMessage } from '../../../../common/warning_message'
 import { Arbitrators } from '../../../common_sections/arbitrators/arbitrators'
+import { WarningMessage } from '../../../common_sections/messages/warning_message'
 import { Outcome, Outcomes } from '../outcomes'
 
 import { CreateScalarMarket } from './create_scalar_market'
