+++ app/src/components/market/sections/market_create/steps/items/funding_and_fee_step.tsx
@@ -47,8 +47,8 @@ import { MarketScale } from '../../../../common/market_scale'
 import { SetAllowance } from '../../../../common/set_allowance'
 import { TradingFeeSelector } from '../../../../common/trading_fee_selector'
 import { VerifiedRow } from '../../../../common/verified_row'
-import { WarningMessage } from '../../../../common/warning_message'
 import { DisplayArbitrator } from '../../../common_sections/arbitrators/display_arbitrator'
+import { WarningMessage } from '../../../common_sections/messages/warning_message'
 import { TransactionDetailsCard } from '../../../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../../../common_sections/transaction/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../../../common_sections/transaction/transaction_details_row'
