+++ app/src/components/market/sections/market_pooling/scalar_market_pool_liquidity.tsx
@@ -58,7 +58,7 @@ import { MarketScale } from '../../common/market_scale'
 import { SetAllowance } from '../../common/set_allowance'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { TokenBalance } from '../../common/token_balance'
-import { WarningMessage } from '../../common/warning_message'
+import { WarningMessage } from '../common_sections/messages/warning_message'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../common_sections/transaction/transaction_details_row'
