+++ app/src/components/market/sections/market_profile/market_status/closed.tsx
@@ -26,10 +26,10 @@ import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
 import { ModalTransactionWrapper } from '../../../../modal'
 import { MarginsButton } from '../../../common/common_styled'
-import MarketResolutionMessage from '../../../common/market_resolution_message'
 import { MarketScale } from '../../../common/market_scale'
 import { MarketTopDetailsClosed } from '../../../common/market_top_details_closed'
 import { ViewCard } from '../../../common/view_card'
+import MarketResolutionMessage from '../../common_sections/messages/market_resolution_message'
 import { OutcomeTable } from '../../common_sections/outcome_table'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
