+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -21,7 +21,7 @@ import { ButtonType } from '../../../../button/button_styling_types'
 import { MarketScale } from '../../../common/market_scale'
 import { MarketTopDetailsOpen } from '../../../common/market_top_details_open'
 import { ViewCard } from '../../../common/view_card'
-import { WarningMessage } from '../../../common/warning_message'
+import { WarningMessage } from '../../common_sections/messages/warning_message'
 import { OutcomeTable } from '../../common_sections/outcome_table'
 import { MarketBondContainer } from '../../market_bond/market_bond_container'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
