+++ app/src/components/market/sections/market_sell/scalar_market_sell.tsx
@@ -45,7 +45,7 @@ import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { MarketScale } from '../../common/market_scale'
 import { PositionSelectionBox } from '../../common/position_selection_box'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
-import { WarningMessage } from '../../common/warning_message'
+import { WarningMessage } from '../common_sections/messages/warning_message'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../common_sections/transaction/transaction_details_row'
